Core.Evaluate

Documentation

normalise :: Context -> Env -> TT Name -> TT Name

normaliseTrace :: Bool -> Context -> Env -> TT Name -> TT Name

normaliseC :: Context -> Env -> TT Name -> TT Name

Normalise fully type checked terms (so, assume all names/let bindings resolved)

normaliseAll :: Context -> Env -> TT Name -> TT Name

rt_simplify :: Context -> Env -> TT Name -> TT Name

Simplify for run-time (i.e. basic inlining)

simplify :: Context -> Env -> TT Name -> TT Name

Like normalise, but we only reduce functions that are marked as okay to inline (and probably shouldn't reduce lets?) 20130908: now only used to reduce for totality checking. Inlining should be done elsewhere.

specialise :: Context -> Env -> [(Name, Int)] -> TT Name -> TT Name

hnf :: Context -> Env -> TT Name -> TT Name

Reduce a term to head normal form

convEq :: Context -> TT Name -> TT Name -> StateT UCs TC Bool

convEq' :: Context -> TT Name -> TT Name -> TC Bool

data Def

A definition is either a simple function (just an expression with a type), a constant, which could be a data or type constructor, an axiom or as an yet undefined function, or an Operator. An Operator is a function which explains how to reduce. A CaseOp is a function defined by a simple case tree

Constructors

Function !Type !Term

TyDecl NameType !Type

Operator Type Int ([Value] -> Maybe Value)

CaseOp CaseInfo !Type ![Either Term (Term, Term)] ![([Name], Term, Term)] !CaseDefs

data CaseInfo

Constructors

CaseInfo

Fields

case_inlinable :: Bool

tc_dictionary :: Bool

data CaseDefs

Constructors

CaseDefs

Fields

cases_totcheck :: !([Name], SC)

cases_compiletime :: !([Name], SC)

cases_inlined :: !([Name], SC)

cases_runtime :: !([Name], SC)

data Accessibility

Constructors

Public

Frozen

Hidden

data Totality

The result of totality checking

Constructors

Total [Int]
well-founded arguments

Productive
productive

Partial PReason

Unchecked

data PReason

Reasons why a function may not be total

Constructors

Other [Name]

Itself

NotCovering

NotPositive

UseUndef Name

BelieveMe

Mutual [Name]

NotProductive

data Context

Contexts used for global definitions and for proof state. They contain universe constraints and existing definitions.

initContext :: Context

The initial empty context

ctxtAlist :: Context -> [(Name, Def)]

Get the definitions from a context

uconstraints :: Context -> [UConstraint]

next_tvar :: Context -> Int

addToCtxt :: Name -> Term -> Type -> Context -> Context

setAccess :: Name -> Accessibility -> Context -> Context

setTotal :: Name -> Totality -> Context -> Context

addCtxtDef :: Name -> Def -> Context -> Context

addTyDecl :: Name -> NameType -> Type -> Context -> Context

addDatatype :: Datatype Name -> Context -> Context

addCasedef :: Name -> CaseInfo -> Bool -> Bool -> Bool -> Bool -> [Either Term (Term, Term)] -> [([Name], Term, Term)] -> [([Name], Term, Term)] -> [([Name], Term, Term)] -> [([Name], Term, Term)] -> Type -> Context -> Context

simplifyCasedef :: Name -> Context -> Context

addOperator :: Name -> Type -> Int -> ([Value] -> Maybe Value) -> Context -> Context

lookupNames :: Name -> Context -> [Name]

lookupTy :: Name -> Context -> [Type]

lookupP :: Name -> Context -> [Term]

lookupDef :: Name -> Context -> [Def]

lookupDefAcc :: Name -> Bool -> Context -> [(Def, Accessibility)]

lookupVal :: Name -> Context -> [Value]

lookupTotal :: Name -> Context -> [Totality]

lookupNameTotal :: Name -> Context -> [(Name, Totality)]

lookupTyEnv :: Name -> Env -> Maybe (Int, Type)

isDConName :: Name -> Context -> Bool

isTConName :: Name -> Context -> Bool

isConName :: Name -> Context -> Bool

isFnName :: Name -> Context -> Bool

data Value

A HOAS representation of values

Constructors

VP NameType Name Value

VV Int

VBind Bool Name (Binder Value) (Value -> Eval Value)

VBLet Int Name Value Value Value

VApp Value Value

VType UExp

VErased

VImpossible

VConstant Const

VProj Value Int

VTmp Int

class Quote a where

Methods

quote :: Int -> a -> Eval (TT Name)

initEval :: EvalState