Core.TT

Description

TT is the core language of Idris. The language has:

Full dependent types

A hierarchy of universes, with cumulativity: Type : Type1, Type1 : Type2, ...

Pattern matching letrec binding

(primitive types defined externally)

Some technical stuff:

Typechecker is kept as simple as possible - no unification, just a checker for incomplete terms.

We have a simple collection of tactics which we use to elaborate source programs with implicit syntax into fully explicit terms.

Documentation

data Option

Constructors

TTypeInTType

CheckConv

data FC

Source location. These are typically produced by the parser getFC

Constructors

FC

Fields

fc_fname :: String

Filename

fc_line :: Int

Line number

fc_column :: Int

Column number

newtype FC'

FC with equality

Constructors

FC'

Fields

unwrapFC :: FC

emptyFC :: FC

Empty source location

fileFC :: String -> FC

Source location with file only

data Err

Constructors

Msg String

InternalMsg String

CantUnify Bool Term Term Err [(Name, Type)] Int

InfiniteUnify Name Term [(Name, Type)]

CantConvert Term Term [(Name, Type)]

UnifyScope Name Name Term [(Name, Type)]

CantInferType String

NonFunctionType Term Term

CantIntroduce Term

NoSuchVariable Name

NoTypeDecl Name

NotInjective Term Term Term

CantResolve Term

CantResolveAlts [String]

IncompleteTerm Term

UniverseError

ProgramLineComment

Inaccessible Name

NonCollapsiblePostulate Name

AlreadyDefined Name

ProofSearchFail Err

NoRewriting Term

At FC Err

Elaborating String Name Err

ProviderError String

LoadingFailed String Err

score :: Err -> Int

data TC a

Constructors

OK a

Error Err

tfail :: Err -> TC a

trun :: FC -> TC a -> TC a

discard :: Monad m => m a -> m ()

showSep :: String -> [String] -> String

pmap :: (t -> t1) -> (t, t) -> (t1, t1)

traceWhen :: Bool -> String -> a -> a

data Name

Names are hierarchies of strings, describing scope (so no danger of duplicate names, but need to be careful on lookup).

Constructors

UN String
User-provided name

NS Name [String]
Root, namespaces

MN Int String
Machine chosen names

NErased
Name of somethng which is never used in scope

SN SpecialName
Decorated function names

data SpecialName

Constructors

WhereN Int Name Name

InstanceN Name [String]

ParentN Name String

MethodN Name

CaseN Name

showCG :: Name -> String

type Ctxt a = Map Name (Map Name a)

Contexts allow us to map names to things. A root name maps to a collection of things in different namespaces with that name.

emptyContext :: Map k a

tcname :: Name -> Bool

implicitable :: Name -> Bool

nsroot :: Name -> Name

addDef :: Name -> a -> Ctxt a -> Ctxt a

lookupCtxtName :: Name -> Ctxt a -> [(Name, a)]

Look up a name in the context, given an optional namespace. The name (n) may itself have a (partial) namespace given.

Rules for resolution:

if an explicit namespace is given, return the names which match it. If none match, return all names.

if the name has has explicit namespace given, return the names which match it and ignore the given namespace.

otherwise, return all names.

lookupCtxt :: Name -> Ctxt a -> [a]

lookupCtxtExact :: Name -> Ctxt a -> [a]

updateDef :: Name -> (a -> a) -> Ctxt a -> Ctxt a

toAlist :: Ctxt a -> [(Name, a)]

addAlist :: Show a => [(Name, a)] -> Ctxt a -> Ctxt a

data NativeTy

Constructors

IT8

IT16

IT32

IT64

data IntTy

Constructors

ITFixed NativeTy

ITNative

ITBig

ITChar

ITVec NativeTy Int

data ArithTy

Constructors

ATInt IntTy

ATFloat

nativeTyWidth :: NativeTy -> Int

intTyWidth :: IntTy -> Int

Deprecated: Non-total function, use nativeTyWidth and appropriate casing

data Const

Constructors

I Int

BI Integer

Fl Double

Ch Char

Str String

B8 Word8

B16 Word16

B32 Word32

B64 Word64

B8V (Vector Word8)

B16V (Vector Word16)

B32V (Vector Word32)

B64V (Vector Word64)

AType ArithTy

StrType

PtrType

VoidType

Forgot

data Raw

Constructors

Var Name

RBind Name (Binder Raw) Raw

RApp Raw Raw

RType

RForce Raw

RConstant Const

data Binder b

All binding forms are represented in a unform fashion.

Constructors

Lam

Fields

binderTy :: b

type annotation for bound variable

Pi

Fields

binderTy :: b

type annotation for bound variable

Let

Fields

binderTy :: b

type annotation for bound variable

binderVal :: b

value for bound variable

NLet

Fields

binderTy :: b

type annotation for bound variable

binderVal :: b

value for bound variable

Hole

Fields

binderTy :: b

type annotation for bound variable

GHole

Fields

envlen :: Int

binderTy :: b

type annotation for bound variable

Guess

Fields

binderTy :: b

type annotation for bound variable

binderVal :: b

value for bound variable

PVar

Fields

binderTy :: b

type annotation for bound variable

PVTy

Fields

binderTy :: b

type annotation for bound variable

fmapMB :: Monad m => (a -> m b) -> Binder a -> m (Binder b)

raw_apply :: Raw -> [Raw] -> Raw

raw_unapply :: Raw -> (Raw, [Raw])

data RawFun

Constructors

RawFun

Fields

rtype :: Raw

rval :: Raw

data RawDatatype

Constructors

RDatatype Name Raw [(Name, Raw)]

data RDef

Constructors

RFunction RawFun

RConst Raw

RData RawDatatype

type RProgram = [(Name, RDef)]

data UExp

Universe expressions for universe checking

Constructors

UVar Int
universe variable

UVal Int
explicit universe level

data UConstraint

Universe constraints

Constructors

ULT UExp UExp
Strictly less than

ULE UExp UExp
Less than or equal to

type UCs = (Int, [UConstraint])

data NameType

Constructors

Bound

Ref

DCon Int Int
Data constructor; Ints are tag and arity

TCon Int Int
Type constructor; Ints are tag and arity

data TT n

Terms in the core language

Constructors

P NameType n (TT n)
named references

V Int
a resolved de Bruijn-indexed variable

Bind n (Binder (TT n)) (TT n)
a binding

App (TT n) (TT n)
function, function type, arg

Constant Const
constant

Proj (TT n) Int
argument projection; runtime only (-1) is a special case for 'subtract one from BI'

Erased
an erased term

Impossible
special case for totality checking

TType UExp
the type of types at some level

class TermSize a where

Methods

termsize :: Name -> a -> Int

type EnvTT n = [(n, Binder (TT n))]

data Datatype n

Constructors

Data

Fields

d_typename :: n

d_typetag :: Int

d_type :: TT n

d_cons :: [(n, TT n)]

A few handy operations on well typed terms:

isInjective :: TT n -> Bool

A term is injective iff it is a data constructor, type constructor, constant, the type Type, pi-binding, or an application of an injective term.

vinstances :: Int -> TT n -> Int

Count the number of instances of a de Bruijn index in a term

instantiate :: TT n -> TT n -> TT n

substV :: TT n -> TT n -> TT n

explicitNames :: TT n -> TT n

pToV :: Eq n => n -> TT n -> TT n

pToV' :: Eq n => n -> Int -> TT n -> TT n

addBinder :: TT n -> TT n

pToVs :: Eq n => [n] -> TT n -> TT n

Convert several names. First in the list comes out as V 0

vToP :: TT n -> TT n

finalise :: Eq n => TT n -> TT n

subst :: Eq n => n -> TT n -> TT n -> TT n

substNames :: Eq n => [(n, TT n)] -> TT n -> TT n

substTerm :: Eq n => TT n -> TT n -> TT n -> TT n

noOccurrence :: Eq n => n -> TT n -> Bool

freeNames :: Eq n => TT n -> [n]

Returns all names used free in the term

arity :: TT n -> Int

Return the arity of a (normalised) type

unApply :: TT n -> (TT n, [TT n])

Deconstruct an application; returns the function and a list of arguments

mkApp :: TT n -> [TT n] -> TT n

forget :: TT Name -> Raw

bindAll :: [(n, Binder (TT n))] -> TT n -> TT n

bindTyArgs :: (TT n -> Binder (TT n)) -> [(n, TT n)] -> TT n -> TT n

getArgTys :: TT n -> [(n, TT n)]

getRetTy :: TT n -> TT n

uniqueName :: Name -> [Name] -> Name

uniqueBinders :: [Name] -> TT Name -> TT Name

nextName :: Name -> Name

type Term = TT Name

type Type = Term

type Env = EnvTT Name

newtype WkEnvTT n

Constructors

Wk (EnvTT n)

type WkEnv = WkEnvTT Name

itBitsName :: NativeTy -> [Char]

showEnv :: (Eq a, Show a) => [(a, Binder (TT a))] -> TT a -> [Char]

showEnvDbg :: (Eq a, Show a) => [(a, Binder (TT a))] -> TT a -> [Char]

prettyEnv :: (Eq a, Show a, Pretty a) => [(a, Binder (TT a))] -> TT a -> Doc

showEnv' :: (Eq a, Show a) => [(a, Binder (TT a))] -> TT a -> Bool -> [Char]

pureTerm :: TT Name -> Bool

Check whether a term has any holes in it - impure if so

weakenTm :: Int -> TT n -> TT n

Weaken a term by adding i to each de Bruijn index (i.e. lift it over i bindings)

weakenEnv :: EnvTT n -> EnvTT n

Weaken an environment so that all the de Bruijn indices are correct according to the latest bound variable

weakenTmEnv :: Int -> EnvTT n -> EnvTT n

orderPats :: Term -> Term