Idris.AbsSyntax

Documentation

getContext :: Idris Context

getObjectFiles :: Idris [FilePath]

addObjectFile :: FilePath -> Idris ()

getLibs :: Idris [String]

addLib :: String -> Idris ()

addHdr :: String -> Idris ()

totcheck :: (FC, Name) -> Idris ()

setFlags :: Name -> [FnOpt] -> Idris ()

setAccessibility :: Name -> Accessibility -> Idris ()

setTotality :: Name -> Totality -> Idris ()

getTotality :: Name -> Idris Totality

getCoercionsTo :: IState -> Type -> [Name]

addToCG :: Name -> CGInfo -> Idris ()

addCoercion :: Name -> Idris ()

addDocStr :: Name -> String -> Idris ()

addToCalledG :: Name -> [Name] -> Idris ()

addInstance :: Bool -> Name -> Name -> Idris ()

addClass :: Name -> ClassInfo -> Idris ()

addIBC :: IBCWrite -> Idris ()

clearIBC :: Idris ()

getHdrs :: Idris [String]

setErrLine :: Int -> Idris ()

clearErr :: Idris ()

getSO :: Idris (Maybe String)

setSO :: Maybe String -> Idris ()

getIState :: Idris IState

putIState :: IState -> Idris ()

getName :: Idris Int

checkUndefined :: FC -> Name -> Idris ()

isUndefined :: FC -> Name -> Idris Bool

setContext :: Context -> Idris ()

updateContext :: (Context -> Context) -> Idris ()

addConstraints :: FC -> (Int, [UConstraint]) -> Idris ()

addDeferred :: [(Name, Type)] -> Idris ()

solveDeferred :: Name -> Idris ()

iputStrLn :: String -> Idris ()

iWarn :: FC -> String -> Idris ()

setLogLevel :: Int -> Idris ()

setCmdLine :: [Opt] -> Idris ()

getDumpDefun :: Idris (Maybe FilePath)

getDumpCases :: Idris (Maybe FilePath)

logLevel :: Idris Int

setErrContext :: Bool -> Idris ()

errContext :: Idris Bool

useREPL :: Idris Bool

setREPL :: Bool -> Idris ()

setTarget :: Target -> Idris ()

target :: Idris Target

setOutputTy :: OutputType -> Idris ()

outputTy :: Idris OutputType

verbose :: Idris Bool

setVerbose :: Bool -> Idris ()

typeInType :: Idris Bool

setTypeInType :: Bool -> Idris ()

coverage :: Idris Bool

setCoverage :: Bool -> Idris ()

setIBCSubDir :: FilePath -> Idris ()

valIBCSubDir :: IState -> Idris FilePath

addImportDir :: FilePath -> Idris ()

setImportDirs :: [FilePath] -> Idris ()

allImportDirs :: IState -> Idris [FilePath]

impShow :: Idris Bool

setImpShow :: Bool -> Idris ()

logLvl :: Int -> String -> Idris ()

cmdOptType :: Opt -> Idris Bool

iLOG :: String -> Idris ()

noErrors :: Idris Bool

setTypeCase :: Bool -> Idris ()

bi :: FC

inferTy :: Name

inferCon :: Name

inferDecl :: PData' PTerm

infTerm :: PTerm -> PTerm

infP :: TT Name

getInferTerm :: Term -> Term

getInferType :: Term -> Term

primNames :: [Name]

unitTy :: Name

unitCon :: Name

unitDecl :: PData' PTerm

falseTy :: Name

falseDecl :: PData' PTerm

pairTy :: Name

pairCon :: Name

pairDecl :: PData' PTerm

eqTy :: Name

eqCon :: Name

eqDecl :: PData' PTerm

sigmaTy :: Name

existsCon :: Name

piBind :: [(Name, PTerm)] -> PTerm -> PTerm

piBindp :: Plicity -> [(Name, PTerm)] -> PTerm -> PTerm

expandParams :: (Name -> Name) -> [(Name, PTerm)] -> [Name] -> [Name] -> PTerm -> PTerm

expandParamsD :: Bool -> IState -> (Name -> Name) -> [(Name, PTerm)] -> [Name] -> PDecl -> PDecl

mapsnd :: (t -> t2) -> (t1, t) -> (t1, t2)

if it's just a type variable or concrete type, do it early (0)

if there's only type variables and injective constructors, do it next (1)

if there's a function type, next (2)

finally, everything else (3)

getPriority :: IState -> PTerm -> Int

addStatics :: Name -> Term -> PTerm -> Idris ()

implicit :: SyntaxInfo -> Name -> PTerm -> Idris PTerm

implicit' :: SyntaxInfo -> [Name] -> Name -> PTerm -> Idris PTerm

implicitise :: SyntaxInfo -> [Name] -> IState -> PTerm -> (PTerm, [PArg])

addImplPat :: IState -> PTerm -> PTerm

addImplBound :: IState -> [Name] -> PTerm -> PTerm

addImplBoundInf :: IState -> [Name] -> [Name] -> PTerm -> PTerm

addImpl :: IState -> PTerm -> PTerm

addImpl' :: Bool -> [Name] -> [Name] -> IState -> PTerm -> PTerm

aiFn :: Bool -> Bool -> IState -> FC -> Name -> [PArg] -> Either Err PTerm

mkPApp :: FC -> Int -> PTerm -> [PArg] -> PTerm

findStatics :: IState -> PTerm -> (PTerm, [Bool])

dumpDecls :: [PDecl] -> String

dumpDecl :: PDecl' PTerm -> [Char]

data EitherErr a b

Constructors

LeftErr a

RightOK b

toEither :: EitherErr a b -> Either a b

matchClause :: IState -> PTerm -> PTerm -> Either (PTerm, PTerm) [(Name, PTerm)]

matchClause' :: Bool -> IState -> PTerm -> PTerm -> Either (PTerm, PTerm) [(Name, PTerm)]

substMatches :: [(Name, PTerm)] -> PTerm -> PTerm

substMatchesShadow :: [(Name, PTerm)] -> [Name] -> PTerm -> PTerm

substMatch :: Name -> PTerm -> PTerm -> PTerm

substMatchShadow :: Name -> [Name] -> PTerm -> PTerm -> PTerm

shadow :: Name -> Name -> PTerm -> PTerm

module Idris.AbsSyntaxTree