Idris.AbsSyntaxTree

Documentation

data IOption

Constructors

IOption

Fields

opt_logLevel :: Int

opt_typecase :: Bool

opt_typeintype :: Bool

opt_coverage :: Bool

opt_showimp :: Bool

opt_errContext :: Bool

opt_repl :: Bool

opt_verbose :: Bool

opt_target :: Target

opt_outputTy :: OutputType

opt_ibcsubdir :: FilePath

opt_importdirs :: [FilePath]

opt_cmdline :: [Opt]

defaultOpts :: IOption

data IState

The global state used in the Idris monad

Constructors

IState

Fields

tt_ctxt :: Context

All the currently defined names and their terms

idris_constraints :: [(UConstraint, FC)]

A list of universe constraints and their corresponding source locations

idris_infixes :: [FixDecl]

Currently defined infix operators

idris_implicits :: Ctxt [PArg]

idris_statics :: Ctxt [Bool]

idris_classes :: Ctxt ClassInfo

idris_dsls :: Ctxt DSL

idris_optimisation :: Ctxt OptInfo

idris_datatypes :: Ctxt TypeInfo

idris_patdefs :: Ctxt [([Name], Term, Term)]

idris_flags :: Ctxt [FnOpt]

idris_callgraph :: Ctxt CGInfo

idris_calledgraph :: Ctxt [Name]

idris_docstrings :: Ctxt String

idris_totcheck :: [(FC, Name)]

idris_log :: String

idris_options :: IOption

idris_name :: Int

idris_metavars :: [Name]

The currently defined but not proven metavariables

idris_coercions :: [Name]

syntax_rules :: [Syntax]

syntax_keywords :: [String]

imported :: [FilePath]

The imported modules

idris_scprims :: [(Name, (Int, PrimFn))]

idris_objs :: [FilePath]

idris_libs :: [String]

idris_hdrs :: [String]

proof_list :: [(Name, [String])]

errLine :: Maybe Int

lastParse :: Maybe Name

indent_stack :: [Int]

brace_stack :: [Maybe Int]

hide_list :: [(Name, Maybe Accessibility)]

default_access :: Accessibility

default_total :: Bool

ibc_write :: [IBCWrite]

compiled_so :: Maybe String

data SizeChange

Constructors

Smaller

Same

Bigger

Unknown

type SCGEntry = (Name, [Maybe (Int, SizeChange)])

data CGInfo

Constructors

CGInfo

Fields

argsdef :: [Name]

calls :: [(Name, [[Name]])]

scg :: [SCGEntry]

argsused :: [Name]

unusedpos :: [Int]

primDefs :: [Name]

data IBCWrite

Constructors

IBCFix FixDecl

IBCImp Name

IBCStatic Name

IBCClass Name

IBCInstance Bool Name Name

IBCDSL Name

IBCData Name

IBCOpt Name

IBCSyntax Syntax

IBCKeyword String

IBCImport FilePath

IBCObj FilePath

IBCLib String

IBCHeader String

IBCAccess Name Accessibility

IBCTotal Name Totality

IBCFlags Name [FnOpt]

IBCCG Name

IBCDoc Name

IBCCoercion Name

IBCDef Name

idrisInit :: IState

type Idris = StateT IState IO

The monad for the main REPL - reading and processing files and updating global state (hence the IO inner monad).

data Target

Constructors

ViaC

ViaJava

ViaNode

ViaJavaScript

Bytecode

data Command

REPL commands

Constructors

Quit

Help

Eval PTerm

Check PTerm

DocStr Name

TotCheck Name

Reload

Load FilePath

ModImport String

Edit

Compile Target String

Execute

ExecVal PTerm

NewCompile String

Metavars

Prove Name

AddProof (Maybe Name)

RmProof Name

ShowProof Name

Proofs

Universes

TTShell

LogLvl Int

Spec PTerm

HNF PTerm

Defn Name

Info Name

Missing Name

Pattelab PTerm

DebugInfo Name

Search PTerm

SetOpt Opt

UnsetOpt Opt

NOP

data Opt

Constructors

Filename String

Ver

Usage

ShowLibs

ShowLibdir

ShowIncs

NoPrelude

NoREPL

OLogging Int

Output String

NewOutput String

TypeCase

TypeInType

DefaultTotal

DefaultPartial

WarnPartial

NoCoverage

ErrContext

ShowImpl

Verbose

IBCSubDir String

ImportDir String

PkgBuild String

PkgInstall String

PkgClean String

WarnOnly

Pkg String

BCAsm String

DumpDefun String

DumpCases String

FOVM String

UseTarget Target

OutputTy OutputType

data Fixity

Constructors

Infixl

Fields

prec :: Int

Infixr

Fields

prec :: Int

InfixN

Fields

prec :: Int

PrefixN

Fields

prec :: Int

data FixDecl

Constructors

Fix Fixity String

data Static

Constructors

Static

Dynamic

data Plicity

Constructors

Imp

Fields

plazy :: Bool

pstatic :: Static

pdocstr :: String

Exp

Fields

plazy :: Bool

pstatic :: Static

pdocstr :: String

Constraint

Fields

plazy :: Bool

pstatic :: Static

pdocstr :: String

TacImp

Fields

plazy :: Bool

pstatic :: Static

pscript :: PTerm

pdocstr :: String

impl :: Plicity

expl :: Plicity

constraint :: Plicity

tacimpl :: PTerm -> Plicity

data FnOpt

Constructors

Inlinable

TotalFn

PartialFn

Coinductive

AssertTotal

TCGen

Implicit

CExport String

Specialise [Name]

type FnOpts = [FnOpt]

inlinable :: FnOpts -> Bool

data PDecl' t

Top-level declarations such as compiler directives, definitions, datatypes and typeclasses.

Constructors

PFix FC Fixity [String]
Fixity declaration

PTy String SyntaxInfo FC FnOpts Name t
Type declaration

PPostulate String SyntaxInfo FC FnOpts Name t
Postulate

PClauses FC FnOpts Name [PClause' t]
Pattern clause

PCAF FC Name t
Top level constant

PData String SyntaxInfo FC Bool (PData' t)
Data declaration. The Bool argument is True for codata.

PParams FC [(Name, t)] [PDecl' t]
Params block

PNamespace String [PDecl' t]
New namespace

PRecord String SyntaxInfo FC Name t String Name t
Record declaration

PClass String SyntaxInfo FC [t] Name [(Name, t)] [PDecl' t]
Type class: arguments are documentation, syntax info, source location, constraints, class name, parameters, method declarations

PInstance SyntaxInfo FC [t] Name [t] t (Maybe Name) [PDecl' t]
Instance declaration: arguments are syntax info, source location, constraints, class name, parameters, full instance type, optional explicit name, and definitions

PDSL Name (DSL' t)
DSL declaration

PSyntax FC Syntax
Syntax definition

PMutual FC [PDecl' t]
Mutual block

PDirective (Idris ())
Compiler directive. The parser inserts the corresponding action in the Idris monad.

data PClause' t

Constructors

PClause FC Name t [t] t [PDecl' t]

PWith FC Name t [t] t [PDecl' t]

PClauseR FC [t] t [PDecl' t]

PWithR FC [t] t [PDecl' t]

data PData' t

Data declaration

Constructors

PDatadecl
Data declaration

Fields

d_name :: Name

The name of the datatype

d_tcon :: t

Type constructor

d_cons :: [(String, Name, t, FC)]

Constructors

PLaterdecl
Placeholder for data whose constructors are defined later

Fields

d_name :: Name

The name of the datatype

d_tcon :: t

Type constructor

type PDecl = PDecl' PTerm

type PData = PData' PTerm

type PClause = PClause' PTerm

declared :: PDecl -> [Name]

tldeclared :: PDecl -> [Name]

defined :: PDecl -> [Name]

updateN :: [(Name, Name)] -> Name -> Name

updateNs :: [(Name, Name)] -> PTerm -> PTerm

data PTerm

High level language terms

Constructors

PQuote Raw

PRef FC Name

PInferRef FC Name
A name to be defined later

PPatvar FC Name

PLam Name PTerm PTerm

PPi Plicity Name PTerm PTerm

PLet Name PTerm PTerm PTerm

PTyped PTerm PTerm
Term with explicit type

PApp FC PTerm [PArg]

PCase FC PTerm [(PTerm, PTerm)]

PTrue FC

PFalse FC

PRefl FC PTerm

PResolveTC FC

PEq FC PTerm PTerm

PPair FC PTerm PTerm

PDPair FC PTerm PTerm PTerm

PAlternative Bool [PTerm]

PHidden PTerm
Irrelevant or hidden pattern

PType

PConstant Const

Placeholder

PDoBlock [PDo]

PIdiom FC PTerm

PReturn FC

PMetavar Name

PProof [PTactic]
Proof script

PTactics [PTactic]
As PProof, but no auto solving

PElabError Err
Error to report on elaboration

PImpossible
Special case for declaring when an LHS can't typecheck

PCoerced PTerm
To mark a coerced argument, so as not to coerce twice

mapPT :: (PTerm -> PTerm) -> PTerm -> PTerm

data PTactic' t

Constructors

Intro [Name]

Intros

Focus Name

Refine Name [Bool]

Rewrite t

LetTac Name t

LetTacTy Name t t

Exact t

Compute

Trivial

Solve

Attack

ProofState

ProofTerm

Undo

Try (PTactic' t) (PTactic' t)

TSeq (PTactic' t) (PTactic' t)

ReflectTac t

GoalType String (PTactic' t)

Qed

Abandon

type PTactic = PTactic' PTerm

data PDo' t

Constructors

DoExp FC t

DoBind FC Name t

DoBindP FC t t

DoLet FC Name t t

DoLetP FC t t

type PDo = PDo' PTerm

data PArg' t

Constructors

PImp

Fields

priority :: Int

lazyarg :: Bool

pname :: Name

getTm :: t

pargdoc :: String

PExp

Fields

priority :: Int

lazyarg :: Bool

getTm :: t

pargdoc :: String

PConstraint

Fields

priority :: Int

lazyarg :: Bool

getTm :: t

pargdoc :: String

PTacImplicit

Fields

priority :: Int

lazyarg :: Bool

pname :: Name

getScript :: t

getTm :: t

pargdoc :: String

pimp :: Name -> t -> PArg' t

pexp :: t -> PArg' t

pconst :: t -> PArg' t

ptacimp :: Name -> t -> t -> PArg' t

type PArg = PArg' PTerm

data ClassInfo

Constructors

CI

Fields

instanceName :: Name

class_methods :: [(Name, (FnOpts, PTerm))]

class_defaults :: [(Name, (Name, PDecl))]

class_params :: [Name]

class_instances :: [Name]

data OptInfo

Constructors

Optimise

Fields

collapsible :: Bool

forceable :: [Int]

recursive :: [Int]

data TypeInfo

Constructors

TI

Fields

con_names :: [Name]

codata :: Bool

param_pos :: [Int]

data DSL' t

Constructors

DSL

Fields

dsl_bind :: t

dsl_return :: t

dsl_apply :: t

dsl_pure :: t

dsl_var :: Maybe t

index_first :: Maybe t

index_next :: Maybe t

dsl_lambda :: Maybe t

dsl_let :: Maybe t

type DSL = DSL' PTerm

data SynContext

Constructors

PatternSyntax

TermSyntax

AnySyntax

data Syntax

Constructors

Rule [SSymbol] PTerm SynContext

data SSymbol

Constructors

Keyword Name

Symbol String

Binding Name

Expr Name

SimpleExpr Name

initDSL :: DSL' PTerm

data SyntaxInfo

Constructors

Syn

Fields

using :: [(Name, PTerm)]

syn_params :: [(Name, PTerm)]

syn_namespace :: [String]

no_imp :: [Name]

decoration :: Name -> Name

inPattern :: Bool

implicitAllowed :: Bool

dsl_info :: DSL

defaultSyntax :: SyntaxInfo

expandNS :: SyntaxInfo -> Name -> Name

showDeclImp :: Bool -> PDecl' PTerm -> [Char]

showCImp :: Bool -> PClause -> String

showDImp :: Bool -> PData -> String

getImps :: [PArg] -> [(Name, PTerm)]

getExps :: [PArg] -> [PTerm]

getConsts :: [PArg] -> [PTerm]

getAll :: [PArg] -> [PTerm]

prettyImp :: Bool -> PTerm -> Doc

showImp :: Bool -> PTerm -> String

getPArity :: PTerm -> Int

allNamesIn :: PTerm -> [Name]

namesIn :: [(Name, PTerm)] -> IState -> PTerm -> [Name]

usedNamesIn :: [Name] -> IState -> PTerm -> [Name]