Idris.ElabTerm

Documentation

data ElabInfo

Constructors

EInfo

Fields

params :: [(Name, PTerm)]

inblock :: Ctxt [Name]

liftname :: Name -> Name

namespace :: Maybe [String]

toplevel :: ElabInfo

build :: IState -> ElabInfo -> Bool -> Name -> PTerm -> ElabD (Term, [(Name, (Int, Maybe Name, Type))], [PDecl])

buildTC :: IState -> ElabInfo -> Bool -> Bool -> Name -> PTerm -> ElabD (Term, [(Name, (Int, Maybe Name, Type))], [PDecl])

elab :: IState -> ElabInfo -> Bool -> Bool -> Name -> PTerm -> ElabD ()

pruneAlt :: [PTerm] -> [PTerm]

pruneByType :: Term -> Context -> [PTerm] -> [PTerm]

findInstances :: IState -> Term -> [Name]

trivial' :: IState -> ElabD ()

proofSearch' :: IState -> Maybe Name -> Name -> [Name] -> ElabD ()

resolveTC :: Int -> Name -> IState -> ElabD ()

collectDeferred :: Maybe Name -> Term -> State [(Name, (Int, Maybe Name, Type))] Term

runTac :: Bool -> IState -> PTactic -> ElabD ()

reflm :: String -> Name

Prefix a name with the Reflection.Language namespace

reify :: IState -> Term -> ElabD PTactic

Reify tactics from their reflected representation

reifyApp :: IState -> Name -> [Term] -> ElabD PTactic

reifyTT :: Term -> ElabD Term

Reify terms from their reflected representation

reifyTTApp :: Name -> [Term] -> ElabD Term

reifyRaw :: Term -> ElabD Raw

Reify raw terms from their reflected representation

reifyRawApp :: Name -> [Term] -> ElabD Raw

reifyTTName :: Term -> ElabD Name

reifyTTNameApp :: Name -> [Term] -> ElabD Name

reifyTTNamespace :: Term -> ElabD [String]

reifyTTNameType :: Term -> ElabD NameType

reifyTTBinder :: (Term -> ElabD a) -> Name -> Term -> ElabD (Binder a)

reifyTTBinderApp :: (Term -> ElabD a) -> Name -> [Term] -> ElabD (Binder a)

reifyTTConst :: Term -> ElabD Const

reifyTTConstApp :: Name -> Term -> ElabD Const

reifyTTUExp :: Term -> ElabD UExp

reflCall :: String -> [Raw] -> Raw

Create a reflected call to a named function/constructor

reflect :: Term -> Raw

Lift a term into its Language.Reflection.TT representation

reflectNameType :: NameType -> Raw

reflectName :: Name -> Raw

reflectBinder :: Binder Term -> Raw

reflectConstant :: Const -> Raw

reflectUExp :: UExp -> Raw

reflectEnv :: Env -> Raw

Reflect the environment of a proof into a List (TTName, Binder TT)

reflectErr :: Err -> Raw

Reflect an error into the internal datatype of Idris -- TODO

envTupleType :: Raw

solveAll :: Elab' aux ()

mkSpecialised :: IState -> FC -> Name -> [PTerm] -> PTerm -> ElabD PTerm

mkSpecDecl :: IState -> Name -> [(PTerm, Bool)] -> PTerm -> ElabD PTerm