Idris.Parser

Main grammar

moduleHeader :: IdrisParser [String]

Parses module definition ModuleHeader ::= 'module' Identifier_t ';'?;

import_ :: IdrisParser String

Parses an import statement Import ::= 'import' Identifier_t ';'?;

prog :: SyntaxInfo -> IdrisParser [PDecl]

Parses program source Prog ::= Decl* EOF;

decl :: SyntaxInfo -> IdrisParser [PDecl]

decl' :: SyntaxInfo -> IdrisParser PDecl

Parses a top-level declaration with possible syntax sugar Decl' ::= Fixity | FunDecl' | Data | Record | SyntaxDecl ;

syntaxDecl :: SyntaxInfo -> IdrisParser PDecl

Parses a syntax extension declaration (and adds the rule to parser state) SyntaxDecl ::= SyntaxRule;

syntaxRule :: SyntaxInfo -> IdrisParser Syntax

syntaxSym :: IdrisParser SSymbol

Parses a syntax symbol (either binding variable, keyword or expression) SyntaxSym ::= '[' Name_t ']' | '{' Name_t '}' | Name_t | StringLiteral_t ;

fnDecl :: SyntaxInfo -> IdrisParser [PDecl]

Parses a function declaration with possible syntax sugar FunDecl ::= FunDecl';

fnDecl' :: SyntaxInfo -> IdrisParser PDecl

fnOpts :: [FnOpt] -> IdrisParser [FnOpt]

postulate :: SyntaxInfo -> IdrisParser PDecl

Parses a postulate

Postulate ::= DocComment_t? postulate FnOpts* Accesibility? FnOpts* FnName TypeSig Terminator ;

using_ :: SyntaxInfo -> IdrisParser [PDecl]

Parses a using declaration

Using ::= using '(' UsingDeclList ')' OpenBlock Decl* CloseBlock ;

params :: SyntaxInfo -> IdrisParser [PDecl]

Parses a parameters declaration

Params ::= parameters '(' TypeDeclList ')' OpenBlock Decl* CloseBlock ;

mutual :: SyntaxInfo -> IdrisParser [PDecl]

Parses a mutual declaration (for mutually recursive functions)

Mutual ::= mutual OpenBlock Decl* CloseBlock ;

namespace :: SyntaxInfo -> IdrisParser [PDecl]

methodsBlock :: SyntaxInfo -> IdrisParser [PDecl]

Parses a methods block (for type classes and instances) MethodsBlock ::= 'where' OpenBlock FnDecl* CloseBlock

class_ :: SyntaxInfo -> IdrisParser [PDecl]

Parses a type class declaration

ClassArgument ::= Name | '(' Name : Expr ')' ;

Class ::= DocComment_t? Accessibility? 'class' ConstraintList? Name ClassArgument* MethodsBlock? ;

instance_ :: SyntaxInfo -> IdrisParser [PDecl]

Parses a type class instance declaration

Instance ::= 'instance' InstanceName? ConstraintList? Name SimpleExpr* MethodsBlock? ;

InstanceName ::= '[' Name ']';

usingDeclList :: SyntaxInfo -> IdrisParser [Using]

Parses a using declaration list UsingDeclList ::= UsingDeclList' | NameList TypeSig ;

UsingDeclList' ::= UsingDecl | UsingDecl ',' UsingDeclList' ;

NameList ::= Name | Name ',' NameList ;

usingDecl :: SyntaxInfo -> IdrisParser Using

Parses a using declaration UsingDecl ::= FnName TypeSig | FnName FnName+ ;

pattern :: SyntaxInfo -> IdrisParser PDecl

Parse a clause with patterns Pattern ::= Clause;

caf :: SyntaxInfo -> IdrisParser PDecl

Parse a constant applicative form declaration CAF ::= 'let' FnName '=' Expr Terminator;

argExpr :: SyntaxInfo -> IdrisParser PTerm

Parse an argument expression ArgExpr ::= HSimpleExpr | ;

rhs :: SyntaxInfo -> Name -> IdrisParser PTerm

Parse a right hand side of a function RHS ::= '=' Expr | ?= RHSName? Expr | impossible ;

RHSName ::= '{' FnName '}';

clause :: SyntaxInfo -> IdrisParser PClause

Parses a function clause RHSOrWithBlock ::= RHS WhereOrTerminator | with SimpleExpr OpenBlock FnDecl+ CloseBlock ; Clause ::= WExpr+ RHSOrWithBlock | SimpleExpr <== FnName RHS WhereOrTerminator | ArgExpr Operator ArgExpr WExpr* RHSOrWithBlock | FnName ConstraintArg* ImplicitOrArgExpr* WExpr* RHSOrWithBlock ; ImplicitOrArgExpr ::= ImplicitArg | ArgExpr; WhereOrTerminator ::= WhereBlock | Terminator;

wExpr :: SyntaxInfo -> IdrisParser PTerm

whereBlock :: Name -> SyntaxInfo -> IdrisParser ([PDecl], [(Name, Name)])

Parses a where block WhereBlock ::= 'where' OpenBlock Decl+ CloseBlock;

codegen_ :: IdrisParser Codegen

Parses a code generation target language name Codegen ::= C | Java | JavaScript | Node | LLVM | Bytecode ;

directive :: SyntaxInfo -> IdrisParser [PDecl]

Parses a compiler directive StringList ::= String | String ',' StringList ;

Directive ::= % Directive';

Directive' ::= lib CodeGen String_t | link CodeGen String_t | flag CodeGen String_t | include CodeGen String_t | hide Name | freeze Name | access Accessibility | 'default' Totality | logging Natural | dynamic StringList | language TypeProviders | language ErrorReflection ;

pLangExt :: IdrisParser LanguageExt

totality :: IdrisParser Bool

Parses a totality Totality ::= partial | total

provider :: SyntaxInfo -> IdrisParser [PDecl]

Parses a type provider Provider ::= % provide '(' FnName TypeSig ')' with Expr;

transform :: SyntaxInfo -> IdrisParser [PDecl]

Parses a transform Transform ::= % transform Expr ==> Expr

Loading and parsing

parseExpr :: IState -> String -> Result PTerm

Parses an expression from input

parseTactic :: IState -> String -> Result PTactic

Parses a tactic from input

parseImports :: FilePath -> String -> Idris ([String], [String], Maybe Delta)

Parse module header and imports

findFC :: Doc -> (FC, String)

There should be a better way of doing this...

parseProg :: SyntaxInfo -> FilePath -> String -> Maybe Delta -> Idris [PDecl]

A program is a list of declarations, possibly with associated documentation strings.

loadModule :: Handle -> FilePath -> Idris String

Load idris module and show error if something wrong happens

loadModule' :: Handle -> FilePath -> Idris String

Load idris module

loadFromIFile :: Handle -> IFileType -> Idris ()

Load idris code from file

loadSource' :: Handle -> Bool -> FilePath -> Idris ()

Load idris source code and show error if something wrong happens

loadSource :: Handle -> Bool -> FilePath -> Idris ()

Load Idris source code

addHides :: [(Name, Maybe Accessibility)] -> Idris ()

Adds names to hide list

module Idris.ParseExpr

module Idris.ParseData

module Idris.ParseHelpers

module Idris.ParseOps