Idris.Providers

Documentation

providerTy :: FC -> PTerm -> PTerm

Wrap a type provider in the type of type providers

getProvided :: TT Name -> Idris (TT Name)

Handle an error, if the type provider returned an error. Otherwise return the provided term.