Idris.Parser

Description

Parse the full Idris language.

Documentation

type TokenParser a = TokenParser a

type IParser = GenParser Char IState

lexer :: TokenParser IState

whiteSpace :: ParsecT String IState Identity ()

lexeme :: ParsecT String IState Identity a -> ParsecT String IState Identity a

symbol :: String -> ParsecT String IState Identity String

natural :: ParsecT String IState Identity Integer

parens :: ParsecT String IState Identity a -> ParsecT String IState Identity a

semi :: ParsecT String IState Identity String

comma :: ParsecT String IState Identity String

identifier :: ParsecT String IState Identity String

reserved :: String -> ParsecT String IState Identity ()

operator :: ParsecT String IState Identity String

reservedOp :: String -> ParsecT String IState Identity ()

integer :: ParsecT String IState Identity Integer

float :: ParsecT String IState Identity Double

strlit :: ParsecT String IState Identity String

chlit :: ParsecT String IState Identity Char

lchar :: Char -> ParsecT String IState Identity Char

fixErrorMsg :: String -> [String] -> String

loadModule :: FilePath -> Idris String

loadSource :: Bool -> FilePath -> Idris ()

addHides :: [(Name, Maybe Accessibility)] -> Idris ()

parseExpr :: IState -> [Char] -> Either ParseError PTerm

parseTac :: IState -> [Char] -> Either ParseError PTactic

parseImports :: FilePath -> String -> Idris ([String], [String], String, SourcePos)

pHeader :: IParser [String]

pushIndent :: IParser ()

lastIndent :: IParser Int

indent :: IParser Int

popIndent :: IParser ()

openBlock :: IParser ()

closeBlock :: IParser ()

pTerminator :: ParsecT String IState Identity ()

pBarTerminator :: ParsecT String IState Identity ()

pKeepTerminator :: ParsecT String IState Identity ()

notEndApp :: ParsecT [Char] IState Identity ()

notEndBlock :: ParsecT [Char] IState Identity ()

pfc :: IParser FC

Use Parsec's internal state to construct a source code position

pImport :: IParser String

parseProg :: SyntaxInfo -> FilePath -> String -> SourcePos -> Idris [PDecl]

A program is a list of declarations, possibly with associated documentation strings.

collect :: [PDecl] -> [PDecl]

Collect PClauses with the same function name

pFullExpr :: SyntaxInfo -> IParser PTerm

pDecl :: SyntaxInfo -> IParser [PDecl]

Parse a top-level declaration

pFunDecl :: SyntaxInfo -> IParser [PDecl]

pDecl' :: SyntaxInfo -> IParser PDecl

pSyntaxDecl :: SyntaxInfo -> IParser PDecl

pSyntaxRule :: SyntaxInfo -> IParser Syntax

pSynSym :: IParser SSymbol

pFunDecl' :: SyntaxInfo -> IParser PDecl

pPostulate :: SyntaxInfo -> IParser PDecl

pUsing :: SyntaxInfo -> IParser [PDecl]

pParams :: SyntaxInfo -> IParser [PDecl]

pMutual :: SyntaxInfo -> IParser [PDecl]

pNamespace :: SyntaxInfo -> IParser [PDecl]

pFixity :: IParser PDecl

fixity :: IParser (Int -> Fixity)

pClass :: SyntaxInfo -> IParser [PDecl]

pInstance :: SyntaxInfo -> IParser [PDecl]

pExpr :: SyntaxInfo -> ParsecT String IState Identity PTerm

pExpr' :: SyntaxInfo -> IParser PTerm

pExtExpr :: SyntaxInfo -> IParser PTerm

pSimpleExtExpr :: SyntaxInfo -> IParser PTerm

pNoExtExpr :: SyntaxInfo -> ParsecT String IState Identity PTerm

pExtensions :: SyntaxInfo -> [Syntax] -> IParser PTerm

data SynMatch

Constructors

SynTm PTerm

SynBind Name

pExt :: SyntaxInfo -> Syntax -> IParser PTerm

pName :: ParsecT [Char] IState Identity Name

pOpFront :: CParser IState Name

Parser for an operator in function position, i.e. enclosed by `()', with an optional namespace.

pfName :: ParsecT [Char] IState Identity Name

pTotality :: IParser Bool

pAccessibility' :: IParser Accessibility

pAccessibility :: IParser (Maybe Accessibility)

pFnOpts :: [FnOpt] -> IParser [FnOpt]

addAcc :: Name -> Maybe Accessibility -> IParser ()

pCaseExpr :: SyntaxInfo -> ParsecT String IState Identity PTerm

pProofExpr :: SyntaxInfo -> ParsecT String IState Identity PTerm

pTacticsExpr :: SyntaxInfo -> ParsecT String IState Identity PTerm

pSimpleExpr :: SyntaxInfo -> ParsecT String IState Identity PTerm

bracketed :: SyntaxInfo -> ParsecT String IState Identity PTerm

pCaseOpt :: SyntaxInfo -> IParser (PTerm, PTerm)

modifyConst :: SyntaxInfo -> FC -> PTerm -> PTerm

pList :: SyntaxInfo -> ParsecT String IState Identity PTerm

pPair :: SyntaxInfo -> ParsecT String IState Identity PTerm

pAlt :: SyntaxInfo -> ParsecT String IState Identity PTerm

pHSimpleExpr :: SyntaxInfo -> ParsecT String IState Identity PTerm

pApp :: SyntaxInfo -> ParsecT String IState Identity PTerm

pArg :: SyntaxInfo -> IParser PArg

pImplicitArg :: SyntaxInfo -> ParsecT String IState Identity (PArg' PTerm)

pConstraintArg :: SyntaxInfo -> ParsecT String IState Identity (PArg' PTerm)

pRecordType :: SyntaxInfo -> ParsecT String IState Identity PTerm

mkType :: Name -> Name

noImp :: SyntaxInfo -> SyntaxInfo

impOK :: SyntaxInfo -> SyntaxInfo

pTSig :: SyntaxInfo -> ParsecT String IState Identity PTerm

pTExpr :: SyntaxInfo -> ParsecT [Char] IState Identity PTerm

pLambda :: SyntaxInfo -> ParsecT String IState Identity PTerm

pLet :: SyntaxInfo -> ParsecT String IState Identity PTerm

pPi :: SyntaxInfo -> ParsecT String IState Identity PTerm

pConstList :: SyntaxInfo -> IParser [PTerm]

tyDeclList :: SyntaxInfo -> ParsecT [Char] IState Identity [(Name, PTerm)]

tyOptDeclList :: SyntaxInfo -> ParsecT [Char] IState Identity [(Name, PTerm)]

bindList :: (t -> t1 -> t2 -> t2) -> [(t, t1)] -> t2 -> t2

pComprehension :: SyntaxInfo -> ParsecT String IState Identity PTerm

pDoBlock :: SyntaxInfo -> ParsecT String IState Identity PTerm

pDo :: SyntaxInfo -> ParsecT String IState Identity (PDo' PTerm)

pIdiom :: SyntaxInfo -> ParsecT String IState Identity PTerm

pConstant :: IParser Const

pStatic :: IParser Static

table :: [FixDecl] -> [[Operator String IState Identity PTerm]]

toTable :: [FixDecl] -> [[Operator String IState Identity PTerm]]

binary :: String -> (FC -> a -> a -> a) -> Assoc -> Operator [Char] IState Identity a

prefix :: String -> (FC -> a -> a) -> Operator String IState Identity a

backtick :: Operator String IState Identity PTerm

accData :: Maybe Accessibility -> Name -> [Name] -> IParser ()

pRecord :: SyntaxInfo -> IParser PDecl

pDataI :: ParsecT String IState Identity Bool

pData :: SyntaxInfo -> IParser PDecl

bindArgs :: [PTerm] -> PTerm -> PTerm

pConstructor :: SyntaxInfo -> IParser (String, Name, PTerm, FC)

pSimpleCon :: SyntaxInfo -> IParser (String, Name, [PTerm], FC)

pDSL :: SyntaxInfo -> IParser PDecl

checkDSL :: DSL -> IParser ()

pOverload :: SyntaxInfo -> IParser (String, PTerm)

pPattern :: SyntaxInfo -> IParser PDecl

pCAF :: SyntaxInfo -> IParser PDecl

pArgExpr :: SyntaxInfo -> ParsecT String IState Identity PTerm

pRHS :: SyntaxInfo -> Name -> IParser PTerm

pClause :: SyntaxInfo -> IParser PClause

pWExpr :: SyntaxInfo -> IParser PTerm

pWhereblock :: Name -> SyntaxInfo -> IParser ([PDecl], [(Name, Name)])

pDirective :: IParser [PDecl]

pTactic :: SyntaxInfo -> IParser PTactic